requirements.txt
# Data Processing & Storage
pyspark==3.4.1
delta-spark==2.4.0
confluent-kafka==2.2.0
boto3==1.28.0  # Để làm việc với MinIO (S3)

# MLOps & Feature Store
feast
mlflow==2.6.0
bentoml==1.1.1

# AI & Vector DB
tensorflow
lancedb
pandas
numpy

# UI & Visualization
streamlit
plotly

dotenv
pytest
